requirements.txt
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudinary==1.21.0
cycler==0.10.0
decorator==4.4.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
idna==2.9
imageio==2.8.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
networkx==2.4
Pillow==7.1.2
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.23.0
scipy==1.4.1
six==1.15.0
SQLAlchemy==1.3.17
tifffile==2020.6.3
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
